web/requirements.txt
psycopg2==2.7.5
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
requests==2.19.1
Flask-Caching==1.4.0
pylibmc==1.5.2
lxml==4.2.4
SQLAlchemy-Utils==0.33.3
